Fix `clamp` function
The clamp(v,a,b) function causes the return value to cycle around the
interval [a,b) based on the value of v.

However prior to this commit, `clamp` was ignoring the third argument
and using the second argument twice in its place.  This resulted in a
division by zero.  Hence `clamp` didn't work at all.

Even if the arguments were treated correctly, `clamp` incorrectly
multiplied rather than added whenever v < a.  This would have produced
bogus results.

Note that as implemented, `clamp` is undefined for b >= a.